Agente de Notas Fiscais/requirements.txt
streamlit
langchain
langchain-experimental
pandas
huggingface_hub
python-dotenv
packaging==23.2
numpy==1.24.4
tabulate
langchain-community

#pip install -r requirements.txt
#streamlit run main.py